requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
cssselect==1.2.0
discord.py==2.3.2
dnspython==2.4.2
fake-useragent==1.2.1
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
multidict==6.0.4
parse==1.19.1
pip==23.2.1
pyee==8.2.2
pymongo==4.4.1
pyppeteer==1.0.2
pyquery==2.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
setuptools==68.0.0
soupsieve==2.4.1
tqdm==4.66.1
urllib3==1.26.16
w3lib==2.1.2
websockets==10.4
wheel==0.38.4
yarl==1.9.2
zipp==3.16.2
